refactor: migrate to TypeScript and Vite
