+++ package.json
@@ -12,11 +12,11 @@
     "url": "https://github.com/distillpub/distill-template/issues"
   },
   "scripts": {
-    "dev": "rollup -c rollup.config.dev.js -w",
+    "dev": "vite",
     "serve": "http-server",
     "test": "mocha",
     "lint": "eslint",
-    "build": "rollup -c rollup.config.prod.js",
+    "build": "vite build --config vite.components.config.ts && vite build --config vite.transforms.config.ts",
     "prepare": "npm run build"
   },
   "repository": {
@@ -25,7 +25,6 @@
   },
   "devDependencies": {
     "@babel/core": "^7.9.0",
-    "@rollup/plugin-babel": "^5.0.0",
     "bibtex-parse-js": "^0.0.23",
     "chai": "^3.5.0",
     "eslint": "^4.19.1",
@@ -36,18 +35,10 @@
     "marked": "^0.8.2",
     "mocha": "^5.2.0",
     "prismjs": "^1.20.0",
-    "rollup": "^2.7.3",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-copy": "^0.2.3",
-    "rollup-plugin-grapher": "^0.2.0",
-    "rollup-plugin-gzip": "^1.4.0",
-    "rollup-plugin-node-resolve": "^2.0.0",
-    "rollup-plugin-serve": "^1.0.1",
-    "rollup-plugin-string": "^2.0.2",
-    "rollup-plugin-uglify": "^1.0.1",
-    "rollup-watch": "^4.3.1",
     "should": "^13.2.3",
-    "source-map-support": "^0.5.19"
+    "source-map-support": "^0.5.19",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.0"
   },
   "dependencies": {
     "@webcomponents/webcomponentsjs": "^1.3.3",
