+++ src/distill-components/distill-footer-template.js
@@ -1,4 +1,4 @@
-import logo from '../assets/distill-logo.svg';
+import logo from '../assets/distill-logo.svg?raw';
 
 export const footerTemplate = `
 <style>
