+++ src/distill-components/distill-header-template.js
@@ -1,4 +1,4 @@
-import logo from '../assets/distill-logo.svg';
+import logo from '../assets/distill-logo.svg?raw';
 
 export const headerTemplate = `
 <style>
